pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-harvester"
version = "1.0.0"
description = "Production-grade email harvesting CLI with pluggable search backends and deterministic tests."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Sagar Biswas" }
]
keywords = ["email", "harvesting", "scraper", "lead-generation", "cli"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "dnspython>=2.6,<3",
  "requests>=2.32,<3",
  "tqdm>=4.66,<5",
]

[project.optional-dependencies]
selenium = [
  "selenium>=4.20,<5",
  "webdriver-manager>=4.0,<5",
]
dev = [
  "build>=1.2,<2",
  "mypy>=1.11,<2",
  "pip-audit>=2.9,<3",
  "pre-commit>=3.8,<4",
  "pytest>=8.3,<9",
  "pytest-cov>=5,<6",
  "ruff>=0.6,<1",
  "types-requests>=2.32,<3",
]

[project.urls]
Homepage = "https://github.com/SagarBiswas-MultiHAT/Email-Scraper"
Repository = "https://github.com/SagarBiswas-MultiHAT/Email-Scraper"
Issues = "https://github.com/SagarBiswas-MultiHAT/Email-Scraper/issues"

[project.scripts]
email-harvester = "email_harvester.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
show_error_codes = true
pretty = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = [
  "bs4.*",
  "dns.*",
  "requests.*",
  "tqdm.*",
  "selenium.*",
  "webdriver_manager.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=src/email_harvester --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]
